refactored to load celery from an AppConfig
