+++ {{cookiecutter.repo_name}}/config/settings/common.py
@@ -257,6 +257,7 @@ LOGGING = {
 }
 {% if cookiecutter.use_celery == "y" %}
 ########## CELERY
+INSTALLED_APPS += ('{{cookiecutter.repo_name}}.taskman.celery.CeleryConfig',)
 # if you are not using the django database broker (e.g. rabbitmq, redis, memcached), you can remove the next line.
 INSTALLED_APPS += ('kombu.transport.django',)
 BROKER_URL = env("CELERY_BROKER_URL", default='django://')
