--- {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/celery.py
@@ -1,26 +1,0 @@
-{% if cookiecutter.use_celery == "y" %}
-from __future__ import absolute_import
-
-import os
-
-from celery import Celery
-
-# set the default Django settings module for the 'celery' program.
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.local")
-
-from django.conf import settings
-
-app = Celery('{{cookiecutter.repo_name}}')
-
-# Using a string here means the worker will not have to
-# pickle the object when using Windows.
-app.config_from_object('django.conf:settings')
-app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
-
-
-@app.task(bind=True)
-def debug_task(self):
-    print('Request: {0!r}'.format(self.request))
-{% else %}
-# use this as a starting point for your project with celery.
-{% endif %}
