+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/taskman/celery.py
@@ -1,0 +1,32 @@
+{% if cookiecutter.use_celery == "y" %}
+from __future__ import absolute_import
+import os
+from celery import Celery
+from django.apps import AppConfig
+from django.conf import settings
+
+if not settings.configured:
+    # set the default Django settings module for the 'celery' program.
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.local")
+
+
+app = Celery('{{cookiecutter.repo_name}}')
+
+
+class CeleryConfig(AppConfig):
+    name = '{{cookiecutter.repo_name}}.taskman'
+    verbose_name = 'Celery Config'
+
+    def ready(self):
+        # Using a string here means the worker will not have to
+        # pickle the object when using Windows.
+        app.config_from_object('django.conf:settings')
+        app.autodiscover_tasks(lambda: settings.INSTALLED_APPS, force=True)
+
+
+@app.task(bind=True)
+def debug_task(self):
+    print('Request: {0!r}'.format(self.request))
+{% else %}
+# use this as a starting point for your project with celery.
+{% endif %}
